requirements.txt
discord.py
python-dotenv
discord-py-slash-command
requests
json
datetime
time
typing